+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Implemented class method [`FixedToVariableCode.from_lengths`](https://komm.dev/ref/FixedToVariableCode/#from_lengths), which constructs the canonical prefix-free code with the given codeword lengths.
 
+### Fixed
+
+- Added input validation in lossless coding classes.
+
 ## v0.32.0 (2026-07-30)
 
 ### Added